% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{all_dims}
\alias{all_dims}
\title{All dims}
\usage{
all_dims()
}
\description{
This function returns an object that can be used when subsetting tensors with
\code{[}. If you are familiar with python,, this is equivalent to the python Ellipsis
\code{...}, (not to be confused with \code{...} in \code{R}).
}
\examples{
\dontrun{
# in python, if x is a numpy array or tensorflow tensor
x[..., i]
# the ellipsis means "expand to match number of dimension of x".
# to translate the above python expression to R, write:
x[all_dims(), i]
}
}
