% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{Configure an Application}
\usage{
config(config = Sys.getenv("R_CONFIG_ACTIVE", unset = "default"),
  file = "config.yml", arguments = commandArgs(TRUE))
}
\arguments{
\item{config}{The configuration to use. Defaults to the active configuration
for the current environment (as specified by the \code{R_CONFIG_ACTIVE} environment
variable), or \code{default} when unset.}

\item{file}{The configuration file to read.}

\item{arguments}{The command line arguments to be parsed.}
}
\value{
A named \R list, mapping configuration keys to values.
}
\description{
Configure a TensorFlow application.
}
\note{
The \code{tensorflow.config.hook} \R option can be set if custom transformations
of the generated configuration file is desired.
}
\examples{
if (!requireNamespace("yaml", quietly = TRUE))
  return()

# generate a dummy configuration
config <- list(
  default = list(
    number   = 42,
    string   = "hello",
    list     = list(42, "hello", list()),
    override = "old"
  )
)

# write configuration to file
yaml <- yaml::as.yaml(config)
file <- file.path(tempdir(), "config.yml")
writeLines(yaml, con = file)

# invoke 'tensorflow::config()' with the generated
# config file, and override a value with command line
# arguments
tensorflow::config(
  config = "default",
  file = file,
  arguments = "--override=new"
)
}
