% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-generics.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate a Model}
\usage{
evaluate(object, ...)
}
\arguments{
\item{object}{An evaluatable \R object.}

\item{...}{Optional arguments passed on to implementing methods.}
}
\description{
Evaluate a model object. See implementations in the
\link[keras:evaluate.keras.engine.training.Model]{keras} and
\link[tfestimators:evaluate.tf_estimator]{tfestimators} packages.
}
\section{Implementations}{

\itemize{
\item \link[keras:evaluate.keras.engine.training.Model]{keras}
\item \link[tfestimators:evaluate.tf_estimator]{tfestimators}
}
}

