% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-generics.R
\name{export_savedmodel}
\alias{export_savedmodel}
\title{Export a Saved Model}
\usage{
export_savedmodel(object, export_dir_base, ...)
}
\arguments{
\item{object}{An \R object.}

\item{export_dir_base}{A string containing a directory in which to export the
SavedModel.}

\item{...}{Optional arguments passed on to implementing methods.}
}
\value{
The path to the exported directory, as a string.
}
\description{
Serialize a model to disk. See implementations in the
\link[keras:export_savedmodel.keras.engine.training.Model]{keras} and
\link[tfestimators:export_savedmodel.tf_estimator]{tfestimators} packages.
}
\section{Implementations}{

\itemize{
\item \link[keras:export_savedmodel.keras.engine.training.Model]{keras}
\item \link[tfestimators:export_savedmodel.tf_estimator]{tfestimators}
}
}

