% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_tensorflow_extras}
\alias{install_tensorflow_extras}
\title{Install additional Python packages alongside TensorFlow}
\usage{
install_tensorflow_extras(packages, conda = "auto")
}
\arguments{
\item{packages}{Python packages to install}

\item{conda}{Path to conda executable (or "auto" to find conda using the PATH
and other conventional install locations). Only used when TensorFlow is
installed within a conda environment.}
}
\description{
Install additional Python packages alongside TensorFlow
}
\details{
This function requires a version of TensorFlow previously installed
via the \code{\link[=install_tensorflow]{install_tensorflow()}} function.

For virtualenv and conda installations, the specified packages will be
installed into the "r-tensorflow" environment. For system installations
on Windows the specified packages will be installed into the system
package library.
}
