% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_modules.R
\docType{data}
\name{np}
\alias{np}
\title{Main NumPy module}
\format{NumPy module}
\usage{
np
}
\description{
Interface to main NumPy  module. Provides access to top level classes
and functions as well as sub-modules.
}
\keyword{datasets}

