% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{parse_arguments}
\alias{parse_arguments}
\title{Parse Command Line Arguments}
\usage{
parse_arguments(arguments = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{arguments}{A vector of command line arguments. Defaults
to the command line arguments received by \R.}
}
\description{
Parse command line arguments of the form `--key=value` and
`--key value`.
}
