% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_extract_backup.R, R/extract.R
\name{py_ellipsis}
\alias{py_ellipsis}
\alias{py_ellipsis}
\title{Python Ellipsis}
\usage{
py_ellipsis()

py_ellipsis()
}
\description{
This function returns a python Ellipsis \code{...}, (not to be confused with \code{...}
in \code{R}). The python ellipsis is commonly used when subsetting numpy
arrays or tensorflow tensors objects with brackets.

This function returns a python Ellipsis \code{...}, (not to be confused with \code{...}
in \code{R}). The python ellipsis is commonly used when subsetting numpy
arrays or tensorflow tensors objects with brackets.
}
\examples{
\dontrun{
# in python, if x is a numpy array or tensorflow tensor
x[..., i]
# the ellipsis means "expand to match number of dimension of x".
# to translate the above python expression to R, write:
x[py_ellipsis(), i]
}
\dontrun{
# in python, if x is a numpy array or tensorflow tensor
x[..., i]
# the ellipsis means "expand to match number of dimension of x".
# to translate the above python expression to R, write:
x[py_ellipsis(), i]
}
}
