% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{import}
\alias{reexports}
\alias{dict}
\alias{reexports}
\alias{tuple}
\alias{reexports}
\alias{iterate}
\alias{reexports}
\alias{\%as\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{reticulate}{\code{\link[reticulate]{import}}, \code{\link[reticulate]{dict}}, \code{\link[reticulate]{tuple}}, \code{\link[reticulate]{iterate}}, \code{\link[reticulate]{\%as\%}}}
}}

