% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{run_dir}
\alias{run_dir}
\title{Run Directory}
\usage{
run_dir(parent_dir = "runs")
}
\arguments{
\item{parent_dir}{Parent to create run directory within.}
}
\value{
Timestamped run directory.
}
\description{
Timestamped directory for storing training/logging data in a separate
location for each training run. Note that if the \code{RUNDIR} environment
variable is availale it will be returned as the run directory.
}
