% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swagger.R
\name{swagger_from_signature_def}
\alias{swagger_from_signature_def}
\title{Create Swagger Definition}
\usage{
swagger_from_signature_def(signature_def)
}
\arguments{
\item{signature_def}{A signature definition exported indirectly through
'export_savedmodel()' or explicitly using 'build_signature_def()'.}
}
\description{
Creates a Swagger definition from a TensorFlow signature.
}
