% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorboard.R
\name{tensorboard}
\alias{tensorboard}
\title{Visualize Training with TensorBoard}
\usage{
tensorboard(logdir, host = "127.0.0.1", port = "auto")
}
\arguments{
\item{logdir}{Directory where training events are logged.}

\item{host}{What host to listen to}
}
\description{
Visualize Training with TensorBoard
}
