% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorflow-package.R
\docType{package}
\name{tensorflow-package}
\alias{tensorflow}
\alias{tensorflow-package}
\title{tensorflow: R Interface to 'TensorFlow'}
\description{
Interface to 'TensorFlow' \url{https://www.tensorflow.org/}, an open source software library for numerical computation using data flow graphs. Nodes in the graph represent mathematical operations, while the graph edges represent the multidimensional data arrays (tensors) communicated between them. The flexible architecture allows you to deploy computation to one or more 'CPUs' or 'GPUs' in a desktop, server, or mobile device with a single 'API'. 'TensorFlow' was originally developed by researchers and engineers working on the Google Brain Team within Google's Machine Intelligence research organization for the purposes of conducting machine learning and deep neural networks research, but the system is general enough to be applicable in a wide variety of other domains as well.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/tensorflow}
  \item Report bugs at \url{https://github.com/rstudio/tensorflow/issues}
}

}
\author{
\strong{Maintainer}: Tomasz Kalinowski \email{tomasz.kalinowski@rstudio.com} [contributor, copyright holder]

Authors:
\itemize{
  \item JJ Allaire [copyright holder]
  \item Yuan Tang \email{terrytangyuan@gmail.com} (\href{https://orcid.org/0000-0001-5243-233X}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Daniel Falbel \email{daniel@rstudio.com} [contributor, copyright holder]
  \item RStudio [copyright holder, funder]
  \item Dirk Eddelbuettel \email{edd@debian.org} [contributor, copyright holder]
  \item Nick Golding \email{nick.golding.research@gmail.com} [contributor, copyright holder]
  \item Google Inc. (Examples and Tutorials) [contributor, copyright holder]
}

}
\keyword{internal}
