% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorflow.R
\docType{data}
\name{tensorflow}
\alias{slim}
\alias{tensorflow}
\alias{tf}
\alias{tflearn}
\title{Import TensorFlow module}
\format{A tensorflow module}
\usage{
tensorflow(module = NULL, silent = FALSE)

tf

tflearn

slim
}
\arguments{
\item{module}{Name of sub-module to import. Defaults to \code{NULL}, which
imports the main tensorflow module.}

\item{silent}{Return \code{NULL} rather than throwing an error
if the specified module cannot be loaded.}
}
\description{
Import a tensorflow module for use in R. Note that the most
commonly used modules (\code{tf}, \code{tflearn}, and \code{slim})
are pre-imported and available for use when the package is loaded.
}
\details{
Three TensorFlow modules are pre-imported and ready for use when the
\pkg{tensorflow} package is loaded:

\describe{
   \item{\code{tf}}{Main TensorFlow module (\href{https://www.tensorflow.org/api_docs/index.html}{\code{tensorflow}}).}
   \item{\code{tflearn}}{Higher-level API for TensorFlow (\href{https://github.com/tflearn/tflearn}{\code{tensorflow.contrib.learn}}).}
   \item{\code{slim}}{Lightweight library for defining, training and
   evaluating complex models in TensorFlow  (\href{https://github.com/tensorflow/tensorflow/tree/master/tensorflow/contrib/slim}{\code{tensorflow.contrib.slim}}).}
}
}
\examples{
\dontrun{
metrics <- tensorflow("contrib.metrics")
reader <- tensorflow("models.rnn.ptb")
}

}
\keyword{datasets}

