% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{tf}
\alias{tf}
\title{Main TensorFlow module}
\format{TensorFlow module}
\usage{
tf
}
\description{
Interface to main TensorFlow  module. Provides access to top level classes
and functions as well as sub-modules (e.g. \code{tf$nn},
\code{tf$contrib$learn}, etc.).
}
\examples{
\dontrun{
library(tensorflow)

hello <- tf$constant('Hello, TensorFlow!')
zeros <- tf$Variable(tf$zeros(shape(1L)))

sess <- tf$Session()
sess$run(tf$global_variables_initializer())

sess$run(hello)
sess$run(zeros)
}
}
\keyword{datasets}
