% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{tf_config}
\alias{tf_config}
\alias{tf_version}
\title{TensorFlow configuration information}
\usage{
tf_config()

tf_version()
}
\value{
List with information on the current configuration of TensorFlow.
You can determine whether TensorFlow was found using the \code{available}
member (other members vary depending on whether \code{available} is \code{TRUE}
or \code{FALSE})
}
\description{
TensorFlow configuration information
}
\keyword{internal}
