% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{tf_gpu_configured}
\alias{tf_gpu_configured}
\title{TensorFlow GPU configuration information}
\usage{
tf_gpu_configured(verbose = TRUE)
}
\arguments{
\item{verbose}{boolean. Whether to show extra GPU info.}
}
\value{
A bool, whether GPU is configured or not, or NA if could not be
determined.
}
\description{
TensorFlow GPU configuration information
}
\keyword{internal}
