% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_options.R
\name{tf_options}
\alias{tf_options}
\title{Options for TensorFlow Routines}
\usage{
tf_options(optimizer = NULL, initialize.session = NULL,
  run.optimizer = NULL)
}
\arguments{
\item{optimizer}{The optimizer to be used.}

\item{initialize.session}{A function used to create an initialize a new
TensorFlow session.}

\item{run.optimizer}{A function used to run the model. Should accept the parameters
\code{session}, \code{cost}, \code{optimizer} and \code{feed.dict}, and repeatedly
run the optimizer until the cost function has been appropriately minimized.}
}
\description{
Used to control the various facets of layers constructed with the
modeling routines included in this package.
}

