% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{tf_probability}
\alias{tf_probability}
\title{TensorFlow Probability Module}
\usage{
tf_probability()
}
\value{
Reference to \href{https://www.tensorflow.org/probability}{TensorFlow Probability}
functions and classes
}
\description{
TensorFlow Probability Module
}
\examples{
\dontrun{
library(tensorflow)
## one time setup:
# reticulate::py_install("tensorflow_probability")
tfp <- tf_probability()
tfp$distributions$Normal(loc = 0, scale = 1)
}

}
