% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{tf_session_with_seed}
\alias{tf_session_with_seed}
\title{Get TensorFlow Session with Random Seed}
\usage{
tf_session_with_seed(seed = 42)
}
\arguments{
\item{seed}{The random seed number}
}
\value{
The TensorFlow Session object with the specified random seed
}
\description{
This function returns a TensorFlow Session with the given seed, particularly useful when
your TensorFlow program needs to give reproducible results. Note that this session forces to use
only single thread since multiple threads are a potential source of non-reproducible results.
For further details, see: https://stackoverflow.com/questions/42022950/which-seeds-have-to-be-set-where-to-realize-100-reproducibility-of-training-res
}
