% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{tf_version}
\alias{tf_version}
\title{Version of Tensorflow}
\usage{
tf_version()
}
\value{
R \code{\link[=numeric_version]{numeric_version()}} object
}
\description{
Version of Tensorflow
}
\keyword{internal}
