% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-generics.R
\name{train_and_evaluate}
\alias{train_and_evaluate}
\title{Simultaneously Train and Evaluate a Model}
\usage{
train_and_evaluate(object, ...)
}
\arguments{
\item{object}{An \R object.}

\item{...}{Optional arguments passed on to implementing methods.}
}
\description{
Train and evaluate a model object. See implementation in the
\link[tfestimators:train_and_evaluate.tf_estimator]{tfestimators} package.
}
\section{Implementations}{

\itemize{
\item \link[tfestimators:train_and_evaluate.tf_estimator]{tfestimators}
}
}

