% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{use_compat}
\alias{use_compat}
\title{Use Compatibility}
\usage{
use_compat(version = c("v1", "v2"))
}
\arguments{
\item{version}{The version to activate. Must be \code{"v1"} or \code{"v2"}}
}
\description{
Enables TensorFlow to run under a different API version for compatibility
with previous versions. For instance, this is useful to run TensorFlow 1.x
code when using TensorFlow 2.x.
}
\examples{
\dontrun{
library(tensorflow)
use_compat("v1")
}

}
