% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{view_savedmodel}
\alias{view_savedmodel}
\title{View a Saved Model}
\usage{
view_savedmodel(export_dir = file.path(getwd(),
  "trained/tensorflow-mnist/1/saved_model.pb"), log_dir = file.path(getwd(),
  "logs"))
}
\arguments{
\item{export_dir}{The path to the exported directory, as a string.}

\item{log_dir}{Directories to scan for training logs. If this is a named
character vector then the specified names will be used as aliases within
TensorBoard.}
}
\value{
The path to the exported directory, as a string.
}
\description{
View a serialized model from disk.
}
