% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{Configure an Application}
\usage{
config(config = Sys.getenv("R_CONFIG_ACTIVE", unset = "default"),
  file = "config.yml", arguments = commandArgs(TRUE))
}
\arguments{
\item{config}{The configuration to use. Defaults to the active configuration
for the current environment (as specified by the \code{R_CONFIG_ACTIVE} environment
variable), or \code{default} when unset.}

\item{file}{The configuration file to read.}

\item{arguments}{The command line arguments to be parsed.}
}
\value{
A named \R list, mapping configuration keys to values.
}
\description{
Configure a TensorFlow application.
}
\note{
The \code{tensorflow.config.hook} \R option can be set if custom transformations
of the generated configuration file is desired.
}
\examples{
# examine an example configuration file provided by tensorflow
file <- system.file("examples/config/config.yml", package = "tensorflow")
file.edit(file)

# read the default configuration
tensorflow::config("default", file = file)

# read the alternate configuration: note that
# the default configuration is inherited, but
# we override the 'string' configuration here
tensorflow::config("alternate", file = file)

# override configuration values using command
# line arguments (normally, these would be
# passed in through the command line invocation
# used to start the process)
tensorflow::config("alternate",
                   file = file,
                   arguments = c("--foo=1"))
}
