% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-generics.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate a Model}
\usage{
evaluate(object, ...)
}
\arguments{
\item{object}{An evaluatable \R object.}

\item{...}{Optional arguments passed on to implementing methods.}
}
\description{
Evaluate a model object. See implementations in the
\link[keras3:evaluate.keras.src.models.model.Model]{keras3} package.
}
\section{Implementations}{

\itemize{
\item \link[keras3:evaluate.keras.src.models.model.Model]{keras3}
}
}

