% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_tensorflow}
\alias{install_tensorflow}
\title{Install TensorFlow and it's dependencies}
\usage{
install_tensorflow(gpu = FALSE)
}
\arguments{
\item{gpu}{Install the GPU version of TensorFlow}
}
\description{
Install TensorFlow and it's dependencies
}
