% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_tensorflow}
\alias{install_tensorflow}
\title{Install TensorFlow and it's dependencies}
\usage{
install_tensorflow(method = c("auto", "virtualenv", "conda", "system"),
  version = "latest", gpu = FALSE, package_url = NULL, conda = "auto")
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows (as this isn't supported by TensorFlow). Note also
that since this command runs without privillege the "system" method is
available only on Windows.}

\item{version}{TensorFlow version to install (must be either "latest" or a
full major.minor.patch specification, e.g. "1.1.0").}

\item{gpu}{Install the GPU version of TensorFlow}

\item{package_url}{URL of the TensorFlow package to install (if not specified
this is determined automatically). Note that if this parameter is provied
then the \code{version} and \code{gpu} parameters are ignored.}

\item{conda}{Path to conda executable (or "auto" to find conda using the PATH
and other conventional install locations).}
}
\description{
Install TensorFlow and it's dependencies
}
