% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_tensorflow}
\alias{install_tensorflow}
\title{Install TensorFlow and its dependencies}
\usage{
install_tensorflow(method = c("auto", "virtualenv", "conda", "system"),
  conda = "auto", version = "default", envname = "r-tensorflow",
  extra_packages = c("keras", "tensorflow-hub"),
  restart_session = TRUE)
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows (as this isn't supported by TensorFlow). Note also
that since this command runs without privillege the "system" method is
available only on Windows.}

\item{conda}{Path to conda executable (or "auto" to find conda using the
PATH and other conventional install locations).}

\item{version}{TensorFlow version to install. Specify "default" to install
the CPU version of the latest release. Specify "gpu" to install the GPU
version of the latest release.

You can also provide a full major.minor.patch specification (e.g. "1.1.0"),
appending "-gpu" if you want the GPU version (e.g. "1.1.0-gpu").

Alternatively, you can provide the full URL to an installer binary (e.g.
for a nightly binary).}

\item{envname}{Name of Python environment to install within}

\item{extra_packages}{Additional Python packages to install along with
TensorFlow.}

\item{restart_session}{Restart R session after installing (note this will
only occur within RStudio).}
}
\description{
Install TensorFlow and its dependencies
}
