% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{parse_arguments}
\alias{parse_arguments}
\title{Parse Command Line Arguments}
\usage{
parse_arguments(arguments = NULL)
}
\arguments{
\item{arguments}{A vector of command line arguments. When
\code{NULL} (the default), the command line arguments received
by the current \R process are used.}
}
\description{
Parse command line arguments of the form \code{--key=value} and
\verb{--key value}. The values are assumed to be valid \code{yaml} and
will be converted using \code{\link[yaml:yaml.load]{yaml::yaml.load()}}.
}
