% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{import}
\alias{dict}
\alias{tuple}
\alias{np_array}
\alias{array_reshape}
\alias{iterate}
\alias{\%as\%}
\alias{use_python}
\alias{use_virtualenv}
\alias{use_condaenv}
\alias{flags}
\alias{flag_numeric}
\alias{flag_integer}
\alias{flag_string}
\alias{flag_boolean}
\alias{run_dir}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{reticulate}{\code{\link[reticulate]{\%as\%}}, \code{\link[reticulate]{array_reshape}}, \code{\link[reticulate]{dict}}, \code{\link[reticulate]{import}}, \code{\link[reticulate]{iterate}}, \code{\link[reticulate]{np_array}}, \code{\link[reticulate]{tuple}}, \code{\link[reticulate]{use_condaenv}}, \code{\link[reticulate]{use_python}}, \code{\link[reticulate]{use_virtualenv}}}

  \item{tfruns}{\code{\link[tfruns]{flag_boolean}}, \code{\link[tfruns]{flag_integer}}, \code{\link[tfruns]{flag_numeric}}, \code{\link[tfruns]{flag_string}}, \code{\link[tfruns]{flags}}, \code{\link[tfruns]{run_dir}}}
}}

