% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{run_dir}
\alias{run_dir}
\alias{latest_run}
\alias{clean_runs}
\title{Run Directory}
\usage{
run_dir(runs_dir = "runs")

latest_run(runs_dir = "runs")

clean_runs(runs_dir = "runs", keep = NULL)
}
\arguments{
\item{runs_dir}{Parent directory for runs}

\item{keep}{Number of most recent runs to keep when cleaning runs.}
}
\value{
\code{run_dir()} retruns a timestamped run directory. \code{latest_run()} is a
convenience function for getting the path to the most recent run
directory.
}
\description{
Timestamped directory for storing training/logging data in a separate
location for each training run. Note that if the \code{RUNDIR} environment
variable is availale it will be returned as the run directory.
}
