% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{run_dir}
\alias{run_dir}
\alias{latest_run}
\alias{list_runs}
\title{Run Directory}
\usage{
run_dir(runs_dir = "runs")

latest_run(runs_dir = "runs")

list_runs(runs_dir = "runs", latest_n = NULL)
}
\arguments{
\item{runs_dir}{Parent directory for runs}

\item{latest_n}{List only the latest n runs}
}
\value{
\code{run_dir()} retruns a timestamped run directory. The
\code{list_runs()} function returns a character vector with the
paths of existing run directories (\code{latest_run()} is a convenience
function for getting the path to the most recent run directory).
}
\description{
Timestamped directory for storing training/logging data in a separate
location for each training run. Note that if the \code{RUNDIR} environment
variable is availale it will be returned as the run directory.
}
