% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorboard.R
\name{tensorboard}
\alias{tensorboard}
\title{TensorBoard Visualization Tool}
\usage{
tensorboard(log_dir = ".", host = "127.0.0.1", port = "auto")
}
\arguments{
\item{log_dir}{Root directory for training logs.}

\item{host}{Host for serving TensorBoard}

\item{port}{Port for serving TensorBoard. If "auto" is specified (the
default) then an unused port will be chosen automatically.}
}
\description{
TensorBoard is a tool inspecting and understanding your TensorFlow runs and
graphs.
}
\details{
When TensorBoard is passed a logdir at startup, it recursively walks
the directory tree rooted at logdir looking for subdirectories that contain
tfevents data. Every time it encounters such a subdirectory, it loads it as
a new run, and the frontend will organize the data accordingly.
}
\seealso{
\url{https://github.com/tensorflow/tensorflow/blob/master/tensorflow/tensorboard/}
}
