% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorboard.R
\name{tensorboard}
\alias{tensorboard}
\alias{unique_log_dir}
\title{TensorBoard Visualization Tool}
\usage{
tensorboard(log_dir = ".", action = c("start", "stop"),
  host = "127.0.0.1", port = "auto", launch_browser = interactive())

unique_log_dir(log_dir = "logs")
}
\arguments{
\item{log_dir}{Root directory for training logs.}

\item{action}{Specify whether to start or stop TensorBoard for the
given \code{log_dir} (TensorBoard will be stopped automatically when
the R session from which it is launchedd is terminated).}

\item{host}{Host for serving TensorBoard}

\item{port}{Port for serving TensorBoard. If "auto" is specified (the
default) then an unused port will be chosen automatically.}

\item{launch_browser}{\code{TRUE} to open a web browser for TensorBoard
after launching.}
}
\value{
URL for browsing TensorBoard (invisibly).
}
\description{
TensorBoard is a tool inspecting and understanding your TensorFlow runs and
graphs.
}
\details{
When TensorBoard is passed a logdir at startup, it recursively walks
the directory tree rooted at logdir looking for subdirectories that contain
tfevents data. Every time it encounters such a subdirectory, it loads it as
a new run, and the frontend will organize the data accordingly.

The TensorBoard process will be automatically destroyed when the R
session in which it is launched exits. You can pass \code{action = "stop"`` to manually terminate a TensorBoard process for a given}log_dir`.
}
\seealso{
\url{https://github.com/tensorflow/tensorflow/blob/master/tensorflow/tensorboard/}
}
