% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tensorflow}
\alias{tensorflow}
\alias{tensorflow-package}
\title{TensorFlow for R}
\description{
\href{https://www.tensorflow.org}{TensorFlow} is an open source software library
for numerical computation using data flow graphs. Nodes in the graph
represent mathematical operations, while the graph edges represent the
multidimensional data arrays (tensors) communicated between them. The
flexible architecture allows you to deploy computation to one or more CPUs or
GPUs in a desktop, server, or mobile device with a single API.
}
\details{
The \href{https://www.tensorflow.org/api_docs/python/tf/all_symbols}{TensorFlow
API} is composed of a set of Python modules that enable constructing and
executing TensorFlow graphs. The tensorflow package provides access to the
complete TensorFlow API from within R.

For additional documentation on the tensorflow package see
\href{https://tensorflow.rstudio.com}{https://tensorflow.rstudio.com}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/tensorflow}
  \item Report bugs at \url{https://github.com/rstudio/tensorflow/issues}
}

}
\author{
\strong{Maintainer}: Tomasz Kalinowski \email{tomasz.kalinowski@posit.co} [contributor, copyright holder]

Authors:
\itemize{
  \item JJ Allaire [copyright holder]
  \item Yuan Tang \email{terrytangyuan@gmail.com} (\href{https://orcid.org/0000-0001-5243-233X}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Daniel Falbel \email{daniel@posit.co} [contributor, copyright holder]
  \item Dirk Eddelbuettel \email{edd@debian.org} [contributor, copyright holder]
  \item Nick Golding \email{nick.golding.research@gmail.com} [contributor, copyright holder]
  \item Google Inc. (Examples and Tutorials) [contributor, copyright holder]
  \item Posit, PBC [copyright holder, funder]
}

}
