% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-generics.R
\name{train}
\alias{train}
\title{Train a Model}
\usage{
train(object, ...)
}
\arguments{
\item{object}{A trainable \R object.}

\item{...}{Optional arguments passed on to implementing methods.}
}
\description{
Train a model object.
}
