% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-generics.R
\name{train}
\alias{train}
\title{(Deprecated) Train a Model}
\usage{
train(object, ...)
}
\arguments{
\item{object}{A trainable \R object.}

\item{...}{Optional arguments passed on to implementing methods.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
}
\description{
Train a model object. See implementation in the
\code{tfestimators::train.tf_estimator()} package.
}
\keyword{internal}
