% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-generics.R
\name{train_and_evaluate}
\alias{train_and_evaluate}
\title{(Deprecated) Simultaneously Train and Evaluate a Model}
\usage{
train_and_evaluate(object, ...)
}
\arguments{
\item{object}{An \R object.}

\item{...}{Optional arguments passed on to implementing methods.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
}
\description{
Train and evaluate a model object. See implementation in the
\code{tfestimators::train_and_evaluate.tf_estimator()} package.
}
\keyword{internal}
