% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{unique_dir}
\alias{unique_dir}
\title{Create a timestamp based unique directory name}
\usage{
unique_dir(parent_dir, prefix = NULL, format = "\%Y-\%m-\%dT\%H-\%M-\%OS3Z")
}
\arguments{
\item{parent_dir}{Parent directory}

\item{prefix}{Character value to prepend to directory name}
}
\value{
Unique directory name
}
\description{
Directory name is a timestamp compatible with filesystem naming restrictions
(e.g. "2017-07-13T16-18-51.569Z").
}
\note{
The name of the directory is based on the current time (in milliseconds
resolution). If the generated directory name already exists then the
function will wait long enough to return a directory with a distinct
timestamp.
}
\seealso{
\code{\link[=run_dir]{run_dir()}}
}
