% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{view_savedmodel}
\alias{view_savedmodel}
\title{View a Saved Model}
\usage{
view_savedmodel(model_dir)
}
\arguments{
\item{model_dir}{The path to the exported model, as a string.}
}
\value{
URL for browsing TensorBoard (invisibly).
}
\description{
View a serialized model from disk.
}
